moved flaghelper to util
